Move isRTL to i18n
